Get formData on submit
So far the upload form would fetch the formData when the "Browse ..." button was clicked. I find it way more logical for the formData to be fetched when the "Upload" button is clicked. My user's would add a file and only then change some additional values in an input field and select box. These values haven't been added to the form data although I was using reactive variables. This should add the expected behaviour for most cases.